create queries folder
