--- server/profileQueries.ts
@@ -1,143 +1,0 @@
-import { prisma } from "./index";
-
-export async function getProfileById(id: number) {
-  const userProfile = await prisma.user.findUniqueOrThrow({
-    where: {
-      id,
-    },
-    include: {
-      friends: true,
-      friendRequests: true,
-      friendRequestsSent: true,
-      friendOf: true,
-      events: true,
-      ownedEvents: true,
-    },
-  });
-  if (userProfile) {
-    return userProfile;
-  } else {
-    console.log("Profile not found.");
-    throw new Error(`Profile '${id}' doesn't exist.`);
-  }
-}
-
-export async function getParticipantById(id: number) {
-  const userProfile = await prisma.user.findUniqueOrThrow({
-    where: {
-      id,
-    },
-    include: {
-      friends: false,
-      friendRequests: false,
-      friendRequestsSent: false,
-      friendOf: false,
-      events: false,
-      ownedEvents: false,
-    },
-  });
-  if (userProfile) {
-    return { pic: userProfile.picture, id };
-  } else {
-    console.log("Profile not found.");
-    throw new Error(`Profile '${id}' doesn't exist.`);
-  }
-}
-
-export async function getProfileByUsername(username: string) {
-  try {
-    const userProfile = await prisma.user.findUnique({
-      where: {
-        username,
-      },
-      include: {
-        friends: true,
-        friendRequests: true,
-        friendRequestsSent: true,
-        friendOf: true,
-        events: true,
-        ownedEvents: true,
-      },
-    });
-    if (userProfile) {
-      return userProfile;
-    }
-  } catch (e) {
-    console.log(e);
-  }
-}
-
-export async function getProfilesByPartialUsername(partialUsername: string) {
-  try {
-    const userProfiles = await prisma.user.findMany({
-      where: {
-        username: {
-          contains: partialUsername,
-          mode: "insensitive",
-        },
-      },
-    });
-    return userProfiles;
-  } catch (e) {
-    console.log(e);
-  }
-}
-
-export async function getFriendsByPartialUsername(
-  partialUsername: string,
-  userId: number
-) {
-  //current user = userId
-  try {
-    const user = await prisma.user.findUnique({
-      where: { id: userId },
-      include: { friends: true },
-    });
-
-    const matchingFriends = user?.friends.filter((friend) =>
-      friend.username.toLowerCase().includes(partialUsername.toLowerCase())
-    );
-    return matchingFriends;
-  } catch (e) {
-    console.log(e);
-  }
-}
-
-export async function getProfileByEmail(email: string) {
-  try {
-    const userProfile = await prisma.user.findUnique({
-      where: {
-        email,
-      },
-    });
-    if (userProfile) {
-      return userProfile;
-    }
-  } catch (e) {
-    console.log(e);
-  }
-}
-
-//Move to userqueries
-
-export async function updateUser(email: string, newPic: string) {
-  const user = await getProfileByEmail(email);
-
-  const updatedUser = await prisma.user.update({
-    where: { email },
-    data: { picture: newPic },
-  });
-
-  return updatedUser;
-}
-
-export async function updateUserName(email: string, newUsername: string) {
-  const user = await getProfileByEmail(email);
-
-  const updatedUser = await prisma.user.update({
-    where: { email },
-    data: { username: newUsername },
-  });
-
-  return updatedUser;
-}
